% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundown.R
\docType{package}
\name{boundown}
\alias{boundown}
\alias{boundown-package}
\title{boundown: A package for creating Masters, and PhD theses
using R Markdown and the format guidelines of
Bogazici University, the Institute of Graduate Studies in Social Sciences}
\description{
Based on thesisdown package by Chester Ismay and using the bookdown package and LaTeX and Word thesis templates from Reed College, this package aims to encourage reproducible research using simple Markdown syntax while embedding all of the R code to produce plots and analyses as well.
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

\author{
\strong{Maintainer}: Serhat Çevikel \email{serhat.cevikel@boun.edu.tr}

}
